+++ CHANGELOG.rst
@@ -5,7 +5,7 @@ UNRELEASED
 ----------
 
   * ADDED: fast int8 to uint8 conversion function.
-  * ADDED: continous AE control (AE_RESET_COUNT).
+  * ADDED: continous AE control (-DCONFIG_APPLY_AE=2).
   * CHANGED: Default MIPI PHY and SHIM frecuency to 150MHz. 
 
 2.0.0